[TM-3335] remove comments and simplify marker files
